+++ packages/vue/src/tabs/tab-content.tsx
@@ -1,12 +1,10 @@
-import { type connect } from '@zag-js/tabs'
+import { type ContentProps } from '@zag-js/tabs'
 import { defineComponent, type PropType } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
 import { getValidChildren, type ComponentWithProps } from '../utils'
 import { useTabsContext } from './tabs-context'
 
-type GetContentProps = Parameters<ReturnType<typeof connect>['getContentProps']>[0]
-
-export type TabContentProps = HTMLArkProps<'div'> & GetContentProps
+export type TabContentProps = HTMLArkProps<'div'> & ContentProps
 
 export const TabContent: ComponentWithProps<TabContentProps> = defineComponent({
   name: 'TabContent',
@@ -20,7 +18,7 @@ export const TabContent: ComponentWithProps<TabContentProps> = defineComponent({
     const api = useTabsContext()
 
     return () => (
-      <ark.div {...api.value.getContentProps({ value: props.value })} {...attrs}>
+      <ark.div {...api.value.getContentProps(props)} {...attrs}>
         {() => getValidChildren(slots)}
       </ark.div>
     )