+++ packages/vue/src/tabs/tab-indicator.tsx
@@ -1,11 +1,11 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useTabsContext } from './tabs-context'
 
 export type TabIndicatorProps = HTMLArkProps<'div'>
 
-export const TabIndicator: ComponentWithProps<TabIndicatorProps> = defineComponent({
+export const TabIndicator = defineComponent({
   name: 'TabIndicator',
   setup(_, { slots, attrs }) {
     const api = useTabsContext()