+++ packages/vue/src/tabs/tab-list.tsx
@@ -1,11 +1,11 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useTabsContext } from './tabs-context'
 
 export type TabListProps = HTMLArkProps<'div'>
 
-export const TabList: ComponentWithProps<TabListProps> = defineComponent({
+export const TabList = defineComponent({
   name: 'TabList',
   setup(_, { slots, attrs }) {
     const api = useTabsContext()