+++ packages/vue/src/tabs/tab-trigger.tsx
@@ -1,13 +1,11 @@
 import { type TriggerProps } from '@zag-js/tabs'
-import { defineComponent, h, type PropType } from 'vue'
-import { useUniqueChild, type ComponentWithProps } from '../utils'
+import { defineComponent, type PropType } from 'vue'
+import { ark, type HTMLArkProps } from '../factory'
+import type { Assign } from '../types'
+import { type ComponentWithProps } from '../utils'
 import { useTabsContext } from './tabs-context'
 
-/** This type is here so that the script 'check-exports' passes
- *  because in Vue we don't pass 'children' as props
- */
-
-export type TabTriggerProps = TriggerProps
+export type TabTriggerProps = Assign<HTMLArkProps<'button'>, TriggerProps>
 
 export const TabTrigger: ComponentWithProps<TabTriggerProps> = defineComponent({
   name: 'TabTrigger',
@@ -22,13 +20,10 @@ export const TabTrigger: ComponentWithProps<TabTriggerProps> = defineComponent({
   },
   setup(props, { slots, attrs }) {
     const api = useTabsContext()
-    return () => {
-      const DefaultSlot = useUniqueChild(slots, 'TabsTrigger')
-
-      return h(DefaultSlot, {
-        ...api.value.getTriggerProps({ value: props.value, disabled: props.disabled }),
-        ...attrs,
-      })
-    }
+    return () => (
+      <ark.button {...api.value.getTriggerProps(props)} {...attrs}>
+        {slots.default?.()}
+      </ark.button>
+    )
   },
 })