+++ packages/vue/src/tabs/use-tabs.ts
@@ -1,37 +1,39 @@
-import { connect, machine } from '@zag-js/tabs'
-import { normalizeProps, useMachine } from '@zag-js/vue'
-import { computed, reactive, type ExtractPropTypes } from 'vue'
+import * as tabs from '@zag-js/tabs'
+import { normalizeProps, useMachine, type PropTypes } from '@zag-js/vue'
+import { computed, type ComputedRef } from 'vue'
 import { useEnvironmentContext } from '../environment'
+import type { Optional } from '../types'
 import { useId } from '../utils'
-import type { TabsContext } from './tabs'
 
-export const useTabs = <T extends ExtractPropTypes<TabsContext>>(
-  emit: CallableFunction,
-  context: T,
-) => {
-  const reactiveContext = reactive(context)
+export type UseTabsProps = Optional<tabs.Context, 'id'> & {
+  modelValue?: tabs.Context['value']
+}
+export type UseTabsReturn = ComputedRef<tabs.Api<PropTypes>>
 
+export const useTabs = (props: UseTabsProps, emit: CallableFunction): UseTabsReturn => {
   const getRootNode = useEnvironmentContext()
+  const context = computed(() => {
+    const { modelValue, ...rest } = props
+    return {
+      ...rest,
+      value: modelValue,
+    }
+  })
 
   const [state, send] = useMachine(
-    machine({
-      ...reactiveContext,
-      id: reactiveContext.id || useId().value,
+    tabs.machine({
+      ...context.value,
+      id: context.value.id ?? useId().value,
       getRootNode,
-      value: reactiveContext.defaultValue,
-      onChange(details) {
-        emit('change', details)
-      },
-      onFocus(details) {
-        emit('focus', details)
-      },
-      onDelete(details) {
-        emit('delete', details)
+      onFocus: (details) => emit('focus', details),
+      onDelete: (details) => emit('delete', details),
+      onChange: (details) => {
+        emit('change', details.value)
+        emit('update:modelValue', details.value)
       },
     }),
+    { context },
   )
 
-  return computed(() => connect(state.value, send, normalizeProps))
+  return computed(() => tabs.connect(state.value, send, normalizeProps))
 }
-
-export type UseTabsReturn = ReturnType<typeof useTabs>